Refactor concept note dashboard components
